Изменит компонент ListOffers
