+++ src/pages/main-page/main-page.tsx
@@ -2,12 +2,9 @@ import { Link } from 'react-router-dom';
 import ListOffers from '../../components/list-offers/list-offers';
 import { CITY_NAMES } from '../../const';
 import { Offer } from '../../types/types';
-import Map from '../../components/map/map';
 import { useState } from 'react';
-import { setCurrentCity, SortOption } from '../../features/sorting-offers-by-cities';
+import { setCurrentCity } from '../../features/sorting-offers-by-cities';
 import { useAppDispatch, useAppSelector } from '../../store/hooks';
-import SortOptions from '../../components/sort-options/sort-options';
-import { getOffersForCurrentCityBySortOption } from '../../utils';
 
 type MainPageProps = {
   stateOffers: Offer[];
@@ -26,12 +23,6 @@ function MainPage({stateOffers}: MainPageProps): JSX.Element {
   const handleCityChange = (city: string) => {
     dispatch(setCurrentCity(city));
   };
-
-  const offersForCurrentCity = stateOffers.filter((item) => item.city.name === currentCity);
-  const currentSortOption: SortOption = useAppSelector((state) => state.rental.sortOption);
-
-  const offersForCurrentCityBySortOption: Offer[] = offersForCurrentCity;
-  getOffersForCurrentCityBySortOption(currentSortOption, offersForCurrentCity, offersForCurrentCityBySortOption);
 
   return (
 
@@ -54,22 +45,12 @@ function MainPage({stateOffers}: MainPageProps): JSX.Element {
             </ul>
           </section>
         </div>
-        <div className="cities">
-          <div className="cities__places-container container">
-            <section className="cities__places places">
-              <h2 className="visually-hidden">Places</h2>
-              <b className="places__found">{offersForCurrentCity.length} place{offersForCurrentCity.length > 1 && 's'} to stay in {currentCity}</b>
-
-              <SortOptions currentSortOption={currentSortOption} />
-
-              <ListOffers offers={offersForCurrentCityBySortOption} onListOfferHover={handleListOfferHover} />
-
-            </section>
-            <div className="cities__right-section">
-              <Map city={offersForCurrentCity[0].city} offers={offersForCurrentCity} selectedOffer={selectedOffer} className='cities'/>
-            </div>
-          </div>
-        </div>
+        <ListOffers
+          currentCity={currentCity}
+          onListOfferHover={handleListOfferHover}
+          selectedOffer={selectedOffer}
+          stateOffers={stateOffers}
+        />
       </main>
     </div>
   );
